pub-744: extract table functionality from subscription actions
