+++ src/test/unit/actions/subscriptionActions/subscription.test.ts
@@ -1,10 +1,8 @@
 import { SubscriptionActions } from '../../../../main/resources/actions/subscriptionActions';
 
-
 const userIdWithSubscriptions = 1;
 const userIdWithoutSubscriptions = 2;
 const subscriptionActions = new SubscriptionActions();
-const subscriptionActionsNoData = new SubscriptionActions();
 const mockedCaseSubscription = {
   name: 'Wyman Inc Dispute',
   reference: 'T20217010',
@@ -41,27 +39,3 @@ describe(`getUserSubscriptions(${userIdWithoutSubscriptions}) with valid user id
   });
 });
 
-describe('generate rows functions without subscriptions', () => {
-  subscriptionActionsNoData.userId = 2;
-  it('generateCaseTableRows should return list of case subscriptions', () => {
-    const caseSubscriptionRows = subscriptionActionsNoData.generateCaseTableRows();
-    expect(caseSubscriptionRows.length).toBe(0);
-  });
-
-  it('generateCourtTableRows should return list of court subscriptions', () => {
-    const courtSubscriptionRows = subscriptionActionsNoData.generateCourtTableRows();
-    expect(courtSubscriptionRows.length).toBe(0);
-  });
-});
-
-describe('generate rows functions with subscriptions', () => {
-  it('generateCaseTableRows should return list of case subscriptions', () => {
-    const caseSubscriptionRows = subscriptionActions.generateCaseTableRows();
-    expect(caseSubscriptionRows.length).toBeGreaterThan(0);
-  });
-
-  it('generateCourtTableRows should return list of court subscriptions', () => {
-    const courtSubscriptionRows = subscriptionActions.generateCourtTableRows();
-    expect(courtSubscriptionRows.length).toBeGreaterThan(0);
-  });
-});
